refactor(hooks): standardize hook file naming conventions
- Renamed hook files to include the '.hook' suffix for consistency across the codebase.
- Updated import statements in various components and pages to reflect the new naming convention.
- Ensured all related hooks are correctly referenced, improving code clarity and maintainability.
