+++ app/(admin)/admin/services/page.tsx
@@ -28,7 +28,7 @@ import {
 } from "@/components/ui/dialog";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Alert, AlertDescription } from "@/components/ui/alert";
-import { useAdminServices } from "@/lib/hooks/use-admin-services";
+import { useAdminServices } from "@/lib/hooks/use-admin-services.hook";
 import type { ServiceWithRelations } from "@/lib/types/service-with-relations.type";
 import type { ApiResponse } from "@/lib/types/api-response.type";
 import {
