+++ lib/hooks/useGoogleIndexing.hook.ts
